Clean up, split model_rng source, avoid unnecessary .cpp file
